ddl: migrate test-infra to testify for part of ddl.testSerialDBSuite
